modified some files

+++ strncmp_koto.c
@@ -15,13 +15,13 @@ int main(void)
 	char s51[]="koto";
 	char s52[]="kotoa";
 
-	for(int i = 0;i<20;i++)
+	for(size_t i = 0;i<20;i++)
 	{
-		printf("%6s vs %6s , n:%2d , ft_strncmp:%4d , original_strncmp:%4d , issame:%d\n",s11,s12,i,ft_strncmp(s11,s12,i),strncmp(s11,s12,i),(ft_strncmp(s11,s12,i)==strncmp(s11,s12,i) ? 1 : 0));
-		printf("%6s vs %6s , n:%2d , ft_strncmp:%4d , original_strncmp:%4d , issame:%d\n",s21,s22,i,ft_strncmp(s21,s22,i),strncmp(s21,s22,i),(ft_strncmp(s21,s22,i)==strncmp(s21,s22,i) ? 1 : 0));
-		printf("%6s vs %6s , n:%2d , ft_strncmp:%4d , original_strncmp:%4d , issame:%d\n",s31,s32,i,ft_strncmp(s31,s32,i),strncmp(s31,s32,i),(ft_strncmp(s31,s32,i)==strncmp(s31,s32,i) ? 1 : 0));
-		printf("%6s vs %6s , n:%2d , ft_strncmp:%4d , original_strncmp:%4d , issame:%d\n",s41,s42,i,ft_strncmp(s41,s42,i),strncmp(s41,s42,i),(ft_strncmp(s41,s42,i)==strncmp(s41,s42,i) ? 1 : 0));
-		printf("%6s vs %6s , n:%2d , ft_strncmp:%4d , original_strncmp:%4d , issame:%d\n",s51,s52,i,ft_strncmp(s51,s52,i),strncmp(s51,s52,i),(ft_strncmp(s51,s52,i)==strncmp(s51,s52,i) ? 1 : 0));
+		printf("%6s vs %6s , n:%2zu , ft_strncmp:%4d , original_strncmp:%4d , issame:%d\n",s11,s12,i,ft_strncmp(s11,s12,i),strncmp(s11,s12,i),(ft_strncmp(s11,s12,i)==strncmp(s11,s12,i) ? 1 : 0));
+		printf("%6s vs %6s , n:%2zu , ft_strncmp:%4d , original_strncmp:%4d , issame:%d\n",s21,s22,i,ft_strncmp(s21,s22,i),strncmp(s21,s22,i),(ft_strncmp(s21,s22,i)==strncmp(s21,s22,i) ? 1 : 0));
+		printf("%6s vs %6s , n:%2zu , ft_strncmp:%4d , original_strncmp:%4d , issame:%d\n",s31,s32,i,ft_strncmp(s31,s32,i),strncmp(s31,s32,i),(ft_strncmp(s31,s32,i)==strncmp(s31,s32,i) ? 1 : 0));
+		printf("%6s vs %6s , n:%2zu , ft_strncmp:%4d , original_strncmp:%4d , issame:%d\n",s41,s42,i,ft_strncmp(s41,s42,i),strncmp(s41,s42,i),(ft_strncmp(s41,s42,i)==strncmp(s41,s42,i) ? 1 : 0));
+		printf("%6s vs %6s , n:%2zu , ft_strncmp:%4d , original_strncmp:%4d , issame:%d\n",s51,s52,i,ft_strncmp(s51,s52,i),strncmp(s51,s52,i),(ft_strncmp(s51,s52,i)==strncmp(s51,s52,i) ? 1 : 0));
 		if(result_flg && (
 			ft_strncmp(s11,s12,i)!=strncmp(s11,s12,i) ||
 			ft_strncmp(s21,s22,i)!=strncmp(s21,s22,i) ||
